rust_analyzer

Module cli

Various batch processing tasks, intended primarily for debugging.

Modules

analysis_stats ğŸ”’

Fully type-check project and print various stats, like the number of type errors.

diagnostics ğŸ”’

Analyze all modules in a project for diagnostics. Exits with a non-zero status code if any errors are found.

flags

Grammar for the command-line arguments.

highlight ğŸ”’

Read Rust code on stdin, print HTML highlighted version to stdout.

lsif ğŸ”’

LSIF (language server index format) generator

parse ğŸ”’

Read Rust code on stdin, print syntax tree on stdout.

prime_caches ğŸ”’

Load the project and run cache priming.

progress_report ğŸ”’

A simple progress bar

run_tests ğŸ”’

Run all tests in a project, similar to cargo test, but using the mir interpreter.

rustc_tests ğŸ”’

Run all tests in a project, similar to cargo test, but using the mir interpreter.

scip ğŸ”’

SCIP generator

ssr ğŸ”’

Applies structured search replace rules from the command line.

symbols ğŸ”’

Read Rust code on stdin, print syntax tree on stdout.

unresolved_references ğŸ”’

Reports references in code that the IDE layer cannot resolve.

Enums

Verbosity

Functions

full_name_of_item ğŸ”’

print_memory_usage ğŸ”’

read_stdin ğŸ”’

report_metric ğŸ”’