rust_analyzer::cli::lsif

Module vendored

Types of Language Server Index Format (LSIF). LSIF is a standard format for language servers or other programming tools to dump their knowledge about a workspace.

Based on https://microsoft.github.io/language-server-protocol/specifications/lsif/0.6.0/specification/

Structs

DeclarationTag

DefinitionTag

Document

EdgeData

EdgeDataMultiIn

Entry

Event

Item

MetaData

PackageInformation

Project

https://github.com/Microsoft/language-server-protocol/blob/master/indexFormat/specification.md#the-project-vertex

RangeBasedDocumentSymbol

ReferenceTag

Repository

ResultSet

https://github.com/Microsoft/language-server-protocol/blob/master/indexFormat/specification.md#result-set

ToolInfo

UnknownTag

Enums

DefinitionResultType

DocumentSymbolOrRangeBasedVec

Edge

Element

Encoding

EventKind

EventScope

ItemKind

LocationOrRangeId

RangeTag

Vertex

Type Aliases

Id