rust_analyzer

Module config

Config used by the language server.

Of particular interest is the feature_flags hash map: while other fields configure the server itself, feature flags are passed into analysis, and tweak things like automatic insertion of () in completions.

Modules

client 🔒

false_or_never 🔒

global 🔒

local 🔒

patch_old_style 🔒

See patch_json_for_outdated_configs

single_or_array 🔒

true_or_always 🔒

workspace 🔒

Macros

_config_data 🔒

_default_str 🔒

_default_val 🔒

_impl_for_config_data 🔒

create_bool_or_string_serde 🔒

Structs

CallInfoConfig

CheckOnSaveTargets 🔒

ClientCommandsConfig

ClientConfigInput 🔒

All fields Option<T>, None representing fields not set in a particular JSON/TOML blob.

ClientDefaultConfigData 🔒

Default config values for this grouping.

ClientInfo 🔒

Config

ConfigChange

ConfigErrors

DefaultConfigData 🔒

DiscoverWorkspaceConfig

DocumentSymbolConfig

Configuration for document symbol search requests.

FilesConfig

FullConfigInput 🔒

All of the config levels, all fields Option<T>, to describe fields that are actually set by some rust-analyzer.toml file or JSON blob. An empty rust-analyzer.toml corresponds to all fields being None.

GlobalConfigInput 🔒

All fields Option<T>, None representing fields not set in a particular JSON/TOML blob.

GlobalDefaultConfigData 🔒

Default config values for this grouping.

GlobalWorkspaceLocalConfigInput 🔒

All of the config levels, all fields Option<T>, to describe fields that are actually set by some rust-analyzer.toml file or JSON blob. An empty rust-analyzer.toml corresponds to all fields being None.

HoverActionsConfig

LensConfig

LocalConfigInput 🔒

All fields Option<T>, None representing fields not set in a particular JSON/TOML blob.

LocalDefaultConfigData 🔒

Default config values for this grouping.

NotificationsConfig

RunnablesConfig

Configuration for runnable items, such as main function or tests.

SnippetDef 🔒

WorkspaceConfigInput 🔒

All fields Option<T>, None representing fields not set in a particular JSON/TOML blob.

WorkspaceDefaultConfigData 🔒

Default config values for this grouping.

WorkspaceLocalConfigInput 🔒

Workspace and local config levels, all fields Option<T>, to describe fields that are actually set by some rust-analyzer.toml file or JSON blob. An empty rust-analyzer.toml corresponds to all fields being None.

WorkspaceSymbolConfig

Configuration for workspace symbol search requests.

Enums

AdjustmentHintsDef 🔒

AdjustmentHintsModeDef 🔒

AnnotationLocation

AutoImportExclusion

AutoImportExclusionType

CallableCompletionDef 🔒

CargoFeaturesDef 🔒

ClosureReturnTypeHintsDef 🔒

ClosureStyle 🔒

ConfigErrorInner

DiscriminantHintsDef 🔒

ExprFillDefaultDef 🔒

FilesWatcher

FilesWatcherDef 🔒

ImportGranularityDef 🔒

ImportPrefixDef 🔒

InvocationStrategy 🔒

LifetimeElisionDef 🔒

LinkedProject

ManifestOrProjectJson 🔒

MaxSubstitutionLength

MemoryLayoutHoverRenderKindDef 🔒

NumProcesses

NumThreads

RatomlFile 🔒

RatomlFileKind

ReborrowHintsDef 🔒

RustfmtConfig

SignatureDetail 🔒

SnippetScopeDef 🔒

TargetDirectory

WorkspaceSymbolSearchKindDef 🔒

WorkspaceSymbolSearchScopeDef 🔒

Functions

deserialize_abs_pathbuf 🔒

doc_comment_to_string 🔒

field_props 🔒

get_field_json 🔒

get_field_toml 🔒

schema 🔒

serialize_abs_pathbuf 🔒

to_title_case 🔒

Translate a field name to a title case string suitable for use in the category names on the vscode settings page.

toml_pointer 🔒

validate_toml_table 🔒

Type Aliases

FxIndexMap 🔒

SchemaField 🔒