rust_analyzer

Module diagnostics

Book keeping for keeping diagnostics easily in sync with the client.

Modules

flycheck_to_proto 🔒

This module provides the functionality needed to convert diagnostics from cargo check json format to the LSP diagnostic format.

Structs

DiagnosticCollection 🔒

DiagnosticsMapConfig

Fix 🔒

PackageFlycheckDiagnostic 🔒

WorkspaceFlycheckDiagnostic 🔒

Enums

NativeDiagnosticsFetchKind 🔒

Functions

are_diagnostics_equal 🔒

convert_diagnostic 🔒

fetch_native_diagnostics 🔒

Type Aliases

CheckFixes 🔒

DiagnosticsGeneration 🔒