rust_analyzer::diagnostics

Module flycheck_to_proto

This module provides the functionality needed to convert diagnostics from cargo check json format to the LSP diagnostic format.

Structs

MappedRustDiagnostic 🔒

SubDiagnostic 🔒

Enums

MappedRustChildDiagnostic 🔒

Functions

clippy_code_description 🔒

diagnostic_related_information 🔒

Converts a secondary Rust span to a LSP related information

diagnostic_severity 🔒

Determines the LSP severity from a diagnostic

is_dummy_macro_file 🔒

Checks whether a file name is from macro invocation and does not refer to an actual file.

location 🔒

Converts a Rust span to a LSP location

map_rust_child_diagnostic 🔒

map_rust_diagnostic_to_lsp 🔒

Converts a Rust root diagnostic to LSP form

position 🔒

primary_location 🔒

Extracts a suitable “primary” location from a rustc diagnostic.

resolve_path 🔒

Resolves paths applying any matching path prefix remappings, and then joining the path to the workspace root.

rustc_code_description 🔒