rust_analyzer::flycheck

Struct FlycheckActor

struct FlycheckActor {

    id: usize,
    generation: usize,
    sender: Sender<FlycheckMessage>,
    config: FlycheckConfig,
    config_json: FlycheckConfigJson,
    manifest_path: Option<AbsPathBuf>,
    ws_target_dir: Option<Utf8PathBuf>,
    root: Arc<AbsPathBuf>,
    sysroot_root: Option<AbsPathBuf>,
    scope: FlycheckScope,
    command_handle: Option<CommandHandle<CheckMessage>>,
    command_receiver: Option<Receiver<CheckMessage>>,
    diagnostics_cleared_for: FxHashSet<PackageSpecifier>,
    diagnostics_received: DiagnosticsReceived,

}

A FlycheckActor is a single check instance of a workspace.

Fields
id: usize

The workspace id of this flycheck instance.

generation: usizesender: Sender<FlycheckMessage>config: FlycheckConfigconfig_json: FlycheckConfigJsonmanifest_path: Option<AbsPathBuf>ws_target_dir: Option<Utf8PathBuf>root: Arc<AbsPathBuf>

Either the workspace root of the workspace we are flychecking, or the project root of the project.

sysroot_root: Option<AbsPathBuf>scope: FlycheckScopecommand_handle: Option<CommandHandle<CheckMessage>>

CargoHandle exists to wrap around the communication needed to be able to run cargo check without blocking. Currently the Rust standard library doesn’t provide a way to read sub-process output without blocking, so we have to wrap sub-processes output handling in a thread and pass messages back over a channel.

command_receiver: Option<Receiver<CheckMessage>>

The receiver side of the channel mentioned above.

diagnostics_cleared_for: FxHashSet<PackageSpecifier>diagnostics_received: DiagnosticsReceived
Implementations

impl FlycheckActor

fn new( id: usize, generation: usize, sender: Sender<FlycheckMessage>, config: FlycheckConfig, config_json: FlycheckConfigJson, sysroot_root: Option<AbsPathBuf>, workspace_root: AbsPathBuf, manifest_path: Option<AbsPathBuf>, ws_target_dir: Option<Utf8PathBuf>, ) -> FlycheckActor

fn report_progress(&self, progress: Progress)

fn next_event(&self, inbox: &Receiver<StateChange>) -> Option<Event>

fn run(self, inbox: Receiver<StateChange>)

fn cancel_check_process(&mut self)

fn clear_diagnostics_state(&mut self)

fn explicit_check_command( &self, scope: &FlycheckScope, saved_file: Option<&AbsPath>, ) -> Option<Command>

fn check_command( &self, scope: &FlycheckScope, saved_file: Option<&AbsPath>, target: Option<Target>, ) -> Option<(Command, FlycheckCommandOrigin)>

Construct a Command object for checking the user’s code. If the user has specified a custom command with placeholders that we cannot fill, return None.

fn send(&self, check_task: FlycheckMessage)