rust_analyzer

Module global_state

The context or environment in which the language server functions. In our server implementation this is know as the WorldState.

Each tick provides an immutable snapshot of the state as WorldSnapshot.

Structs

FetchBuildDataResponse 🔒

FetchWorkspaceRequest 🔒

FetchWorkspaceResponse 🔒

GlobalState 🔒

GlobalState is the primary mutable state of the language server

GlobalStateSnapshot 🔒

An immutable snapshot of the world’s state at a point in time.

Handle 🔒

MiniCoreRustAnalyzerInternalOnly 🔒

Functions

file_id_to_url 🔒

url_to_file_id 🔒

Returns None if the file was excluded.

vfs_path_to_file_id 🔒

Returns None if the file was excluded.

Type Aliases

ReqHandler 🔒

ReqQueue 🔒