rust_analyzer::lsp

Module ext

rust-analyzer extensions to the LSP.

Structs

AnalyzerStatusParams

CargoRunnableArgs

ChangeTestStateParams

ClientCommandOptions

CodeAction

CodeActionData

CodeLensResolveData

Information about CodeLens, that is to be resolved.

CommandLink

CommandLinkGroup

CompletionImport

CompletionResolveData

CrateInfoResult

DiscoverTestParams

DiscoverTestResults

EvaluatePredicateParams

EvaluatePredicateResult

ExpandMacroParams

ExpandedMacro

ExternalDocsPair

FetchDependencyListParams

FetchDependencyListResult

GetFailedObligationsParams

Hover

HoverParams

InlayHintResolveData

InternalTestingFetchConfigParams

JoinLinesParams

MatchingBraceParams

MemoryLayoutNode

MoveItemParams

OpenCargoTomlParams

RecursiveMemoryLayout

RunFlycheckParams

RunTestParams

Runnable

RunnablesParams

ServerStatusParams

ShellRunnableArgs

SnippetTextDocumentEdit

SnippetTextEdit

SnippetWorkspaceEdit

SsrParams

TestInfo

TestItem

ViewCrateGraphParams

ViewItemTreeParams

ViewSyntaxTreeParams

WorkspaceSymbolParams

Enums

AbortRunTestNotification

AnalyzerStatusRequest

AppendOutputToRunTestNotification

CancelFlycheckNotification

ChangeTestStateNotification

ChildModulesRequest

ClearFlycheckNotification

CodeActionRequest

CodeActionResolveRequest

CodeLensResolveDataKind

DiscoverTestRequest

DiscoveredTestsNotification

DocumentOnTypeFormattingRequest

The document on type formatting request is sent from the client to the server to format parts of the document during typing. This is almost same as [lsp_types::DocumentOnTypeFormattingRequest], but the result has SnippetTextEdit in it instead of TextEdit.

EndRunTestNotification

EvaluatePredicateRequest

ExpandMacroRequest

ExternalDocsRequest

ExternalDocsResponse

FetchDependencyListRequest

GetFailedObligationsRequest

Health

HoverRequest

InternalTestingFetchConfigOption

InternalTestingFetchConfigRequest

InternalTestingFetchConfigResponse

InterpretFunctionRequest

JoinLinesRequest

MatchingBraceRequest

MemoryUsageRequest

MoveItemDirection

MoveItemRequest

OnEnterRequest

OpenCargoTomlRequest

OpenServerLogsNotification

ParentModuleRequest

PositionOrRange

PredicateEvaluationStatus

RebuildProcMacrosRequest

RelatedTestsRequest

ReloadWorkspaceRequest

RunFlycheckNotification

RunTestRequest

RunnableArgs

RunnablesRequest

ServerStatusNotification

SnippetDocumentChangeOperation

SsrRequest

TestItemKind

TestState

ViewCrateGraphRequest

ViewFileTextRequest

ViewHirRequest

ViewItemTreeRequest

ViewMirRequest

ViewRecursiveMemoryLayoutRequest

ViewSyntaxTreeRequest

WorkspaceSymbolRequest

WorkspaceSymbolSearchKind

WorkspaceSymbolSearchScope