Module rust_analyzer::lsp_ext

rust-analyzer extensions to the LSP.

Structs

AnalyzerStatusParams

CargoRunnable

ClientCommandOptions

CodeAction

CodeActionData

CommandLink

CommandLinkGroup

CompletionImport

CompletionResolveData

ExpandMacroParams

ExpandedMacro

Hover

HoverParams

InlayHintResolveData

InlayHintsParams

JoinLinesParams

MatchingBraceParams

MoveItemParams

OpenCargoTomlParams

RunFlycheckParams

Runnable

RunnablesParams

ServerStatusParams

SnippetTextDocumentEdit

SnippetTextEdit

SnippetWorkspaceEdit

SsrParams

SyntaxTreeParams

TestInfo

ViewCrateGraphParams

ViewItemTreeParams

WorkspaceSymbolParams

Enums

AnalyzerStatus

CancelFlycheck

ClearFlycheck

CodeActionRequest

CodeActionResolveRequest

ExpandMacro

ExternalDocs

Health

HoverRequest

JoinLines

MatchingBrace

MemoryUsage

MoveItem

MoveItemDirection

OnEnter

OnTypeFormatting

The document on type formatting request is sent from the client to the server to format parts of the document during typing. This is almost same as lsp_types::request::OnTypeFormatting, but the result has SnippetTextEdit in it instead of TextEdit.

OpenCargoToml

OpenServerLogs

ParentModule

PositionOrRange

RelatedTests

ReloadWorkspace

RunFlycheck

RunnableKind

Runnables

ServerStatusNotification

ShuffleCrateGraph

SnippetDocumentChangeOperation

Ssr

SyntaxTree

ViewCrateGraph

ViewFileText

ViewHir

ViewItemTree

WorkspaceSymbol

WorkspaceSymbolSearchKind

WorkspaceSymbolSearchScope

Functions

supports_utf8