rust_analyzer

Module test_runner

This module provides the functionality needed to run cargo test in a background thread and report the result of each test in a channel.

Structs

CargoTestHandle ğŸ”’

CargoTestMessage ğŸ”’

CargoTestOutputParser ğŸ”’

TestTarget ğŸ”’

Enums

CargoTestOutput ğŸ”’

TestState ğŸ”’