rust_analyzer
Module tracing

Re-exports

pub use config::Config;

Modules

config

Simple logger that logs either to stderr or to a file, using tracing_subscriber filter syntax and tracing_appender for non blocking output.

hprof

Consumer of tracing data, which prints a hierarchical profile.

json

A [tracing_subscriber::layer::Layer] that exports new-line delinated JSON.