Crate rust_analyzer

Implementation of the LSP for rust-analyzer.

This crate takes Rust-specific analysis results from ide and translates into LSP types.

It also is the root of all state. world module defines the bulk of the state, and main_loop module defines the rules for modifying it.

The cli submodule implements some batch-processing analysis, primarily as a debugging aid.

Modules

cli

Various batch processing tasks, intended primarily for debugging.

config

Config used by the language server.

lsp

Custom LSP definitions and protocol conversions.

tracing

Constants

MINIMUM_SUPPORTED_TOOLCHAIN_VERSION

Any toolchain less than this version will likely not work with rust-analyzer built from this revision.

Functions

from_json

main_loop

server_capabilities

version

Returns information about cargo’s version.

ws_to_crate_graph