Crate rust_analyzer

Implementation of the LSP for rust-analyzer.

This crate takes Rust-specific analysis results from ide and translates into LSP types.

It also is the root of all state. world module defines the bulk of the state, and main_loop module defines the rules for modifying it.

The cli submodule implements some batch-processing analysis, primarily as a debugging aid.

Modules

cli
Various batch processing tasks, intended primarily for debugging.

config
Config used by the language server.

lsp_ext
rust-analyzer extensions to the LSP.

Functions

from_json

main_loop

server_capabilities

Type Definitions

Error

Result