rust_analyzer

Module config

Config used by the language server.

Of particular interest is the feature_flags hash map: while other fields configure the server itself, feature flags are passed into analysis, and tweak things like automatic insertion of () in completions.

Structs

CallInfoConfig

ClientCommandsConfig

Config

ConfigChange

ConfigErrors

DiscoverWorkspaceConfig

DocumentSymbolConfig

Configuration for document symbol search requests.

FilesConfig

HoverActionsConfig

LensConfig

NotificationsConfig

RunnablesConfig

Configuration for runnable items, such as main function or tests.

WorkspaceSymbolConfig

Configuration for workspace symbol search requests.

Enums

AnnotationLocation

AutoImportExclusion

AutoImportExclusionType

ConfigErrorInner

FilesWatcher

LinkedProject

MaxSubstitutionLength

NumThreads

RatomlFileKind

RustfmtConfig

TargetDirectory