Module rust_analyzer::config

Config used by the language server.

We currently get this config from initialize LSP request, which is not the best way to do it, but was the simplest thing we could implement.

Of particular interest is the feature_flags hash map: while other fields configure the server itself, feature flags are passed into analysis, and tweak things like automatic insertion of () in completions.

Structs

CallInfoConfig

ClientCommandsConfig

Config

ConfigUpdateError

FilesConfig

HoverActionsConfig

LensConfig

NotificationsConfig

RunnablesConfig

Configuration for runnable items, such as main function or tests.

WorkspaceSymbolConfig

Configuration for workspace symbol search requests.

Enums

FilesWatcher

LinkedProject

RustfmtConfig