rust_analyzer::config

Struct Config

pub struct Config {

    discovered_projects_from_filesystem: Vec<ProjectManifest>,
    discovered_projects_from_command: Vec<ProjectJsonFromCommand>,
    workspace_roots: Vec<AbsPathBuf>,
    caps: ClientCapabilities,
    root_path: AbsPathBuf,
    snippets: Vec<Snippet>,
    client_info: Option<ClientInfo>,
    default_config: &'static DefaultConfigData,
    client_config: (FullConfigInput, ConfigErrors),
    user_config: Option<(GlobalWorkspaceLocalConfigInput, ConfigErrors)>,
    ratoml_file: FxHashMap<SourceRootId, (RatomlFile, ConfigErrors)>,
    source_root_parent_map: Arc<FxHashMap<SourceRootId, SourceRootId>>,
    validation_errors: ConfigErrors,
    detached_files: Vec<AbsPathBuf>,

}

Fields
discovered_projects_from_filesystem: Vec<ProjectManifest>

Projects that have a Cargo.toml or a rust-project.json in a parent directory, so we can discover them by walking the file system.

discovered_projects_from_command: Vec<ProjectJsonFromCommand>

Projects whose configuration was generated by a command configured in discoverConfig.

workspace_roots: Vec<AbsPathBuf>

The workspace roots as registered by the LSP client

caps: ClientCapabilitiesroot_path: AbsPathBuf

The LSP root path, deprecated in favor of workspace_roots

snippets: Vec<Snippet>client_info: Option<ClientInfo>default_config: &'static DefaultConfigDataclient_config: (FullConfigInput, ConfigErrors)

Config node that obtains its initial value during the server initialization and by receiving a lsp_types::notification::DidChangeConfiguration.

user_config: Option<(GlobalWorkspaceLocalConfigInput, ConfigErrors)>

Config node whose values apply to every Rust project.

ratoml_file: FxHashMap<SourceRootId, (RatomlFile, ConfigErrors)>source_root_parent_map: Arc<FxHashMap<SourceRootId, SourceRootId>>

Clone of the value that is stored inside a GlobalState.

validation_errors: ConfigErrors

Use case : It is an error to have an empty value for check_command. Since it is a global command at the moment, its final value can only be determined by traversing through global configs and the client config. However the non-null value constraint is config level agnostic, so this requires an independent error storage

detached_files: Vec<AbsPathBuf>
Implementations

impl Config

fn cachePriming_enable(&self) -> &bool

fn cachePriming_numThreads(&self) -> &NumThreads

fn completion_snippets_custom( &self, ) -> &IndexMap<String, SnippetDef, FxBuildHasher>

fn files_exclude(&self) -> &Vec<Utf8PathBuf>

fn gotoImplementations_filterAdjacentDerives(&self) -> &bool

fn highlightRelated_branchExitPoints_enable(&self) -> &bool

fn highlightRelated_breakPoints_enable(&self) -> &bool

fn highlightRelated_closureCaptures_enable(&self) -> &bool

fn highlightRelated_exitPoints_enable(&self) -> &bool

fn highlightRelated_references_enable(&self) -> &bool

fn highlightRelated_yieldPoints_enable(&self) -> &bool

fn hover_actions_debug_enable(&self) -> &bool

fn hover_actions_enable(&self) -> &bool

fn hover_actions_gotoTypeDef_enable(&self) -> &bool

fn hover_actions_implementations_enable(&self) -> &bool

fn hover_actions_references_enable(&self) -> &bool

fn hover_actions_run_enable(&self) -> &bool

fn hover_actions_updateTest_enable(&self) -> &bool

fn hover_documentation_enable(&self) -> &bool

fn hover_documentation_keywords_enable(&self) -> &bool

fn hover_dropGlue_enable(&self) -> &bool

fn hover_links_enable(&self) -> &bool

fn hover_maxSubstitutionLength(&self) -> &Option<MaxSubstitutionLength>

fn hover_memoryLayout_alignment( &self, ) -> &Option<MemoryLayoutHoverRenderKindDef>

fn hover_memoryLayout_enable(&self) -> &bool

fn hover_memoryLayout_niches(&self) -> &Option<bool>

fn hover_memoryLayout_offset(&self) -> &Option<MemoryLayoutHoverRenderKindDef>

fn hover_memoryLayout_padding(&self) -> &Option<MemoryLayoutHoverRenderKindDef>

fn hover_memoryLayout_size(&self) -> &Option<MemoryLayoutHoverRenderKindDef>

fn hover_show_enumVariants(&self) -> &Option<usize>

fn hover_show_fields(&self) -> &Option<usize>

fn hover_show_traitAssocItems(&self) -> &Option<usize>

fn inlayHints_bindingModeHints_enable(&self) -> &bool

fn inlayHints_chainingHints_enable(&self) -> &bool

fn inlayHints_closingBraceHints_enable(&self) -> &bool

fn inlayHints_closingBraceHints_minLines(&self) -> &usize

fn inlayHints_closureCaptureHints_enable(&self) -> &bool

fn inlayHints_closureReturnTypeHints_enable(&self) -> &ClosureReturnTypeHintsDef

fn inlayHints_closureStyle(&self) -> &ClosureStyle

fn inlayHints_discriminantHints_enable(&self) -> &DiscriminantHintsDef

fn inlayHints_expressionAdjustmentHints_disableReborrows(&self) -> &bool

fn inlayHints_expressionAdjustmentHints_enable(&self) -> &AdjustmentHintsDef

fn inlayHints_expressionAdjustmentHints_hideOutsideUnsafe(&self) -> &bool

fn inlayHints_expressionAdjustmentHints_mode(&self) -> &AdjustmentHintsModeDef

fn inlayHints_genericParameterHints_const_enable(&self) -> &bool

fn inlayHints_genericParameterHints_lifetime_enable(&self) -> &bool

fn inlayHints_genericParameterHints_type_enable(&self) -> &bool

fn inlayHints_implicitDrops_enable(&self) -> &bool

fn inlayHints_implicitSizedBoundHints_enable(&self) -> &bool

fn inlayHints_impliedDynTraitHints_enable(&self) -> &bool

fn inlayHints_lifetimeElisionHints_enable(&self) -> &LifetimeElisionDef

fn inlayHints_lifetimeElisionHints_useParameterNames(&self) -> &bool

fn inlayHints_maxLength(&self) -> &Option<usize>

fn inlayHints_parameterHints_enable(&self) -> &bool

fn inlayHints_parameterHints_missingArguments_enable(&self) -> &bool

fn inlayHints_rangeExclusiveHints_enable(&self) -> &bool

fn inlayHints_reborrowHints_enable(&self) -> &ReborrowHintsDef

fn inlayHints_renderColons(&self) -> &bool

fn inlayHints_typeHints_enable(&self) -> &bool

fn inlayHints_typeHints_hideClosureInitialization(&self) -> &bool

fn inlayHints_typeHints_hideClosureParameter(&self) -> &bool

fn inlayHints_typeHints_hideInferredTypes(&self) -> &bool

fn inlayHints_typeHints_hideNamedConstructor(&self) -> &bool

fn interpret_tests(&self) -> &bool

fn joinLines_joinAssignments(&self) -> &bool

fn joinLines_joinElseIf(&self) -> &bool

fn joinLines_removeTrailingComma(&self) -> &bool

fn joinLines_unwrapTrivialBlock(&self) -> &bool

fn lens_debug_enable(&self) -> &bool

fn lens_enable(&self) -> &bool

fn lens_implementations_enable(&self) -> &bool

fn lens_location(&self) -> &AnnotationLocation

fn lens_references_adt_enable(&self) -> &bool

fn lens_references_enumVariant_enable(&self) -> &bool

fn lens_references_method_enable(&self) -> &bool

fn lens_references_trait_enable(&self) -> &bool

fn lens_run_enable(&self) -> &bool

fn lens_updateTest_enable(&self) -> &bool

fn linkedProjects(&self) -> &Vec<ManifestOrProjectJson>

fn lru_capacity(&self) -> &Option<u16>

fn lru_query_capacities(&self) -> &FxHashMap<Box<str>, u16>

fn notifications_cargoTomlNotFound(&self) -> &bool

fn numThreads(&self) -> &Option<NumThreads>

fn procMacro_attributes_enable(&self) -> &bool

fn procMacro_enable(&self) -> &bool

fn procMacro_processes(&self) -> &NumProcesses

fn procMacro_server(&self) -> &Option<Utf8PathBuf>

fn profiling_memoryProfile(&self) -> &Option<Utf8PathBuf>

fn references_excludeImports(&self) -> &bool

fn references_excludeTests(&self) -> &bool

fn semanticHighlighting_comments_enable(&self) -> &bool

fn semanticHighlighting_doc_comment_inject_enable(&self) -> &bool

fn semanticHighlighting_nonStandardTokens(&self) -> &bool

fn semanticHighlighting_operator_enable(&self) -> &bool

fn semanticHighlighting_operator_specialization_enable(&self) -> &bool

fn semanticHighlighting_punctuation_enable(&self) -> &bool

fn semanticHighlighting_punctuation_separate_macro_bang(&self) -> &bool

fn semanticHighlighting_punctuation_specialization_enable(&self) -> &bool

fn semanticHighlighting_strings_enable(&self) -> &bool

fn signatureInfo_detail(&self) -> &SignatureDetail

fn signatureInfo_documentation_enable(&self) -> &bool

fn typing_triggerChars(&self) -> &Option<String>

fn workspace_discoverConfig(&self) -> &Option<DiscoverWorkspaceConfig>

impl Config

fn assist_emitMustUse(&self, source_root: Option<SourceRootId>) -> &bool

fn assist_expressionFillDefault( &self, source_root: Option<SourceRootId>, ) -> &ExprFillDefaultDef

fn assist_preferSelf(&self, source_root: Option<SourceRootId>) -> &bool

fn assist_termSearch_borrowcheck( &self, source_root: Option<SourceRootId>, ) -> &bool

fn assist_termSearch_fuel(&self, source_root: Option<SourceRootId>) -> &usize

fn completion_addSemicolonToUnit( &self, source_root: Option<SourceRootId>, ) -> &bool

fn completion_autoAwait_enable( &self, source_root: Option<SourceRootId>, ) -> &bool

fn completion_autoIter_enable(&self, source_root: Option<SourceRootId>) -> &bool

fn completion_autoimport_enable( &self, source_root: Option<SourceRootId>, ) -> &bool

fn completion_autoimport_exclude( &self, source_root: Option<SourceRootId>, ) -> &Vec<AutoImportExclusion>

fn completion_autoself_enable(&self, source_root: Option<SourceRootId>) -> &bool

fn completion_callable_snippets( &self, source_root: Option<SourceRootId>, ) -> &CallableCompletionDef

fn completion_excludeTraits( &self, source_root: Option<SourceRootId>, ) -> &Vec<String>

fn completion_fullFunctionSignatures_enable( &self, source_root: Option<SourceRootId>, ) -> &bool

fn completion_hideDeprecated(&self, source_root: Option<SourceRootId>) -> &bool

fn completion_limit(&self, source_root: Option<SourceRootId>) -> &Option<usize>

fn completion_postfix_enable(&self, source_root: Option<SourceRootId>) -> &bool

fn completion_privateEditable_enable( &self, source_root: Option<SourceRootId>, ) -> &bool

fn completion_termSearch_enable( &self, source_root: Option<SourceRootId>, ) -> &bool

fn completion_termSearch_fuel( &self, source_root: Option<SourceRootId>, ) -> &usize

fn diagnostics_disabled( &self, source_root: Option<SourceRootId>, ) -> &FxHashSet<String>

fn diagnostics_enable(&self, source_root: Option<SourceRootId>) -> &bool

fn diagnostics_experimental_enable( &self, source_root: Option<SourceRootId>, ) -> &bool

fn diagnostics_remapPrefix( &self, source_root: Option<SourceRootId>, ) -> &FxHashMap<String, String>

fn diagnostics_styleLints_enable( &self, source_root: Option<SourceRootId>, ) -> &bool

fn diagnostics_warningsAsHint( &self, source_root: Option<SourceRootId>, ) -> &Vec<String>

fn diagnostics_warningsAsInfo( &self, source_root: Option<SourceRootId>, ) -> &Vec<String>

fn imports_granularity_enforce( &self, source_root: Option<SourceRootId>, ) -> &bool

fn imports_granularity_group( &self, source_root: Option<SourceRootId>, ) -> &ImportGranularityDef

fn imports_group_enable(&self, source_root: Option<SourceRootId>) -> &bool

fn imports_merge_glob(&self, source_root: Option<SourceRootId>) -> &bool

fn imports_preferNoStd(&self, source_root: Option<SourceRootId>) -> &bool

fn imports_preferPrelude(&self, source_root: Option<SourceRootId>) -> &bool

fn imports_prefix(&self, source_root: Option<SourceRootId>) -> &ImportPrefixDef

fn imports_prefixExternPrelude( &self, source_root: Option<SourceRootId>, ) -> &bool

fn rename_showConflicts(&self, source_root: Option<SourceRootId>) -> &bool

impl Config

fn cargo_allTargets(&self, source_root: Option<SourceRootId>) -> &bool

fn cargo_autoreload(&self, source_root: Option<SourceRootId>) -> &bool

fn cargo_buildScripts_enable(&self, source_root: Option<SourceRootId>) -> &bool

fn cargo_buildScripts_invocationStrategy( &self, source_root: Option<SourceRootId>, ) -> &InvocationStrategy

fn cargo_buildScripts_overrideCommand( &self, source_root: Option<SourceRootId>, ) -> &Option<Vec<String>>

fn cargo_buildScripts_rebuildOnSave( &self, source_root: Option<SourceRootId>, ) -> &bool

fn cargo_buildScripts_useRustcWrapper( &self, source_root: Option<SourceRootId>, ) -> &bool

fn cargo_cfgs(&self, source_root: Option<SourceRootId>) -> &Vec<String>

fn cargo_extraArgs(&self, source_root: Option<SourceRootId>) -> &Vec<String>

fn cargo_extraEnv( &self, source_root: Option<SourceRootId>, ) -> &FxHashMap<String, Option<String>>

fn cargo_features(&self, source_root: Option<SourceRootId>) -> &CargoFeaturesDef

fn cargo_noDefaultFeatures(&self, source_root: Option<SourceRootId>) -> &bool

fn cargo_noDeps(&self, source_root: Option<SourceRootId>) -> &bool

fn cargo_sysroot(&self, source_root: Option<SourceRootId>) -> &Option<String>

fn cargo_sysrootSrc(&self, source_root: Option<SourceRootId>) -> &Option<String>

fn cargo_target(&self, source_root: Option<SourceRootId>) -> &Option<String>

fn cargo_targetDir( &self, source_root: Option<SourceRootId>, ) -> &Option<TargetDirectory>

fn cfg_setTest(&self, source_root: Option<SourceRootId>) -> &bool

fn checkOnSave(&self, source_root: Option<SourceRootId>) -> &bool

fn check_allTargets(&self, source_root: Option<SourceRootId>) -> &Option<bool>

fn check_command(&self, source_root: Option<SourceRootId>) -> &String

fn check_extraArgs(&self, source_root: Option<SourceRootId>) -> &Vec<String>

fn check_extraEnv( &self, source_root: Option<SourceRootId>, ) -> &FxHashMap<String, Option<String>>

fn check_features( &self, source_root: Option<SourceRootId>, ) -> &Option<CargoFeaturesDef>

fn check_ignore(&self, source_root: Option<SourceRootId>) -> &FxHashSet<String>

fn check_invocationStrategy( &self, source_root: Option<SourceRootId>, ) -> &InvocationStrategy

fn check_noDefaultFeatures( &self, source_root: Option<SourceRootId>, ) -> &Option<bool>

fn check_overrideCommand( &self, source_root: Option<SourceRootId>, ) -> &Option<Vec<String>>

fn check_targets( &self, source_root: Option<SourceRootId>, ) -> &Option<CheckOnSaveTargets>

fn check_workspace(&self, source_root: Option<SourceRootId>) -> &bool

fn document_symbol_search_excludeLocals( &self, source_root: Option<SourceRootId>, ) -> &bool

fn procMacro_ignored( &self, source_root: Option<SourceRootId>, ) -> &FxHashMap<Box<str>, Box<[Box<str>]>>

fn runnables_bench_command(&self, source_root: Option<SourceRootId>) -> &String

fn runnables_bench_overrideCommand( &self, source_root: Option<SourceRootId>, ) -> &Option<Vec<String>>

fn runnables_command( &self, source_root: Option<SourceRootId>, ) -> &Option<String>

fn runnables_doctest_overrideCommand( &self, source_root: Option<SourceRootId>, ) -> &Option<Vec<String>>

fn runnables_extraArgs(&self, source_root: Option<SourceRootId>) -> &Vec<String>

fn runnables_extraTestBinaryArgs( &self, source_root: Option<SourceRootId>, ) -> &Vec<String>

fn runnables_test_command(&self, source_root: Option<SourceRootId>) -> &String

fn runnables_test_overrideCommand( &self, source_root: Option<SourceRootId>, ) -> &Option<Vec<String>>

fn rustc_source(&self, source_root: Option<SourceRootId>) -> &Option<String>

fn rustfmt_extraArgs(&self, source_root: Option<SourceRootId>) -> &Vec<String>

fn rustfmt_overrideCommand( &self, source_root: Option<SourceRootId>, ) -> &Option<Vec<String>>

fn rustfmt_rangeFormatting_enable( &self, source_root: Option<SourceRootId>, ) -> &bool

fn vfs_extraIncludes(&self, source_root: Option<SourceRootId>) -> &Vec<String>

fn workspace_symbol_search_excludeImports( &self, source_root: Option<SourceRootId>, ) -> &bool

fn workspace_symbol_search_kind( &self, source_root: Option<SourceRootId>, ) -> &WorkspaceSymbolSearchKindDef

fn workspace_symbol_search_limit( &self, source_root: Option<SourceRootId>, ) -> &usize

fn workspace_symbol_search_scope( &self, source_root: Option<SourceRootId>, ) -> &WorkspaceSymbolSearchScopeDef

impl Config

fn files_watcher(&self) -> &FilesWatcherDef

impl Config

pub fn user_config_dir_path() -> Option<AbsPathBuf>

Path to the user configuration dir. This can be seen as a generic way to define what would be $XDG_CONFIG_HOME/rust-analyzer in Linux.

pub fn same_source_root_parent_map( &self, other: &Arc<FxHashMap<SourceRootId, SourceRootId>>, ) -> bool

fn apply_change_with_sink(&self, change: ConfigChange) -> (Config, bool)

Changes made to client and global configurations will partially not be reflected even after .apply_change() was called. The return tuple’s bool component signals whether the GlobalState should call its update_configuration() method.

pub fn apply_change(&self, change: ConfigChange) -> (Config, ConfigErrors, bool)

Given change this generates a new Config, thereby collecting errors of type ConfigError. If there are changes that have global/client level effect, the last component of the return type will be set to true, which should be used by the GlobalState to update itself.

pub fn add_discovered_project_from_command( &mut self, data: ProjectJsonData, buildfile: AbsPathBuf, )

pub fn workspace_roots(&self) -> &[AbsPathBuf]

impl Config

pub fn new( root_path: AbsPathBuf, caps: ClientCapabilities, workspace_roots: Vec<AbsPathBuf>, client_info: Option<ClientInfo>, ) -> Self

pub fn rediscover_workspaces(&mut self)

pub fn remove_workspace(&mut self, path: &AbsPath)

pub fn add_workspaces(&mut self, paths: impl Iterator<Item = AbsPathBuf>)

pub fn json_schema() -> Value

pub fn root_path(&self) -> &AbsPathBuf

pub fn caps(&self) -> &ClientCapabilities

pub fn assist(&self, source_root: Option<SourceRootId>) -> AssistConfig

pub fn rename(&self, source_root: Option<SourceRootId>) -> RenameConfig

pub fn call_hierarchy<'a>( &self, minicore: MiniCore<'a>, ) -> CallHierarchyConfig<'a>

pub fn completion<'a>( &'a self, source_root: Option<SourceRootId>, minicore: MiniCore<'a>, ) -> CompletionConfig<'a>

pub fn completion_hide_deprecated(&self) -> bool

pub fn detached_files(&self) -> &Vec<AbsPathBuf>

pub fn diagnostics( &self, source_root: Option<SourceRootId>, ) -> DiagnosticsConfig

pub fn diagnostic_fixes( &self, source_root: Option<SourceRootId>, ) -> DiagnosticsConfig

pub fn expand_proc_attr_macros(&self) -> bool

pub fn highlight_related( &self, _source_root: Option<SourceRootId>, ) -> HighlightRelatedConfig

pub fn hover_actions(&self) -> HoverActionsConfig

pub fn hover<'a>(&self, minicore: MiniCore<'a>) -> HoverConfig<'a>

pub fn goto_definition<'a>( &self, minicore: MiniCore<'a>, ) -> GotoDefinitionConfig<'a>

pub fn inlay_hints<'a>(&self, minicore: MiniCore<'a>) -> InlayHintsConfig<'a>

fn insert_use_config( &self, source_root: Option<SourceRootId>, ) -> InsertUseConfig

pub fn join_lines(&self) -> JoinLinesConfig

pub fn highlighting_non_standard_tokens(&self) -> bool

pub fn highlighting_config<'a>( &self, minicore: MiniCore<'a>, ) -> HighlightConfig<'a>

pub fn has_linked_projects(&self) -> bool

pub fn linked_manifests(&self) -> impl Iterator<Item = &Utf8Path> + '_

pub fn has_linked_project_jsons(&self) -> bool

pub fn discover_workspace_config(&self) -> Option<&DiscoverWorkspaceConfig>

fn discovered_projects(&self) -> Vec<ManifestOrProjectJson>

pub fn linked_or_discovered_projects(&self) -> Vec<LinkedProject>

pub fn prefill_caches(&self) -> bool

pub fn publish_diagnostics(&self, source_root: Option<SourceRootId>) -> bool

pub fn diagnostics_map( &self, source_root: Option<SourceRootId>, ) -> DiagnosticsMapConfig

pub fn extra_args(&self, source_root: Option<SourceRootId>) -> &Vec<String>

pub fn extra_env( &self, source_root: Option<SourceRootId>, ) -> &FxHashMap<String, Option<String>>

pub fn check_extra_args(&self, source_root: Option<SourceRootId>) -> Vec<String>

pub fn check_extra_env( &self, source_root: Option<SourceRootId>, ) -> FxHashMap<String, Option<String>>

pub fn lru_parse_query_capacity(&self) -> Option<u16>

pub fn lru_query_capacities_config(&self) -> Option<&FxHashMap<Box<str>, u16>>

pub fn proc_macro_srv(&self) -> Option<AbsPathBuf>

pub fn dhat_output_file(&self) -> Option<AbsPathBuf>

pub fn ignored_proc_macros( &self, source_root: Option<SourceRootId>, ) -> &FxHashMap<Box<str>, Box<[Box<str>]>>

pub fn expand_proc_macros(&self) -> bool

pub fn files(&self) -> FilesConfig

pub fn excluded(&self) -> impl Iterator<Item = AbsPathBuf> + use<'_>

pub fn notifications(&self) -> NotificationsConfig

pub fn cargo_autoreload_config(&self, source_root: Option<SourceRootId>) -> bool

pub fn run_build_scripts(&self, source_root: Option<SourceRootId>) -> bool

pub fn cargo(&self, source_root: Option<SourceRootId>) -> CargoConfig

pub fn cfg_set_test(&self, source_root: Option<SourceRootId>) -> bool

pub(crate) fn completion_snippets_default() -> IndexMap<String, SnippetDef, FxBuildHasher>

pub fn rustfmt(&self, source_root_id: Option<SourceRootId>) -> RustfmtConfig

pub fn flycheck_workspace(&self, source_root: Option<SourceRootId>) -> bool

pub(crate) fn cargo_test_options( &self, source_root: Option<SourceRootId>, ) -> CargoOptions

pub(crate) fn flycheck( &self, source_root: Option<SourceRootId>, ) -> FlycheckConfig

fn target_dir_from_config( &self, source_root: Option<SourceRootId>, ) -> TargetDirectoryConfig

pub fn check_on_save(&self, source_root: Option<SourceRootId>) -> bool

pub fn script_rebuild_on_save(&self, source_root: Option<SourceRootId>) -> bool

pub fn runnables(&self, source_root: Option<SourceRootId>) -> RunnablesConfig

pub fn find_all_refs_exclude_imports(&self) -> bool

pub fn find_all_refs_exclude_tests(&self) -> bool

pub fn snippet_cap(&self) -> Option<SnippetCap>

pub fn call_info(&self) -> CallInfoConfig

pub fn lens(&self) -> LensConfig

pub fn goto_implementation(&self) -> GotoImplementationConfig

pub fn document_symbol( &self, source_root: Option<SourceRootId>, ) -> DocumentSymbolConfig

pub fn workspace_symbol( &self, source_root: Option<SourceRootId>, ) -> WorkspaceSymbolConfig

pub fn client_commands(&self) -> ClientCommandsConfig

pub fn prime_caches_num_threads(&self) -> usize

pub fn proc_macro_num_processes(&self) -> usize

pub fn main_loop_num_threads(&self) -> usize

pub fn typing_trigger_chars(&self) -> &str

pub fn visual_studio_code_version(&self) -> Option<&Version>

pub fn client_is_neovim(&self) -> bool