rust_analyzer::lsp

Module ext

rust-analyzer extensions to the LSP.

Structs

AnalyzerStatusParams

CargoRunnableArgs

ChangeTestStateParams

ClientCommandOptions

CodeAction

CodeActionData

CodeLensResolveData

Information about CodeLens, that is to be resolved.

CommandLink

CommandLinkGroup

CompletionImport

CompletionResolveData

CrateInfoResult

DiscoverTestParams

DiscoverTestResults

ExpandMacroParams

ExpandedMacro

ExternalDocsPair

FetchDependencyListParams

FetchDependencyListResult

GetFailedObligationsParams

Hover

HoverParams

InlayHintResolveData

InternalTestingFetchConfigParams

JoinLinesParams

MatchingBraceParams

MemoryLayoutNode

MoveItemParams

OpenCargoTomlParams

RecursiveMemoryLayout

RunFlycheckParams

RunTestParams

Runnable

RunnablesParams

ServerStatusParams

ShellRunnableArgs

SnippetTextDocumentEdit

SnippetTextEdit

SnippetWorkspaceEdit

SsrParams

TestInfo

TestItem

ViewCrateGraphParams

ViewItemTreeParams

ViewSyntaxTreeParams

WorkspaceSymbolParams

Enums

AbortRunTest

AnalyzerStatus

AppendOutputToRunTest

CancelFlycheck

ChangeTestState

ChildModules

ClearFlycheck

CodeActionRequest

CodeActionResolveRequest

CodeLensResolveDataKind

DiscoverTest

DiscoveredTests

EndRunTest

ExpandMacro

ExternalDocs

ExternalDocsResponse

FetchDependencyList

GetFailedObligations

Health

HoverRequest

InternalTestingFetchConfig

InternalTestingFetchConfigOption

InternalTestingFetchConfigResponse

InterpretFunction

JoinLines

MatchingBrace

MemoryUsage

MoveItem

MoveItemDirection

OnEnter

OnTypeFormatting

The document on type formatting request is sent from the client to the server to format parts of the document during typing. This is almost same as lsp_types::request::OnTypeFormatting, but the result has SnippetTextEdit in it instead of TextEdit.

OpenCargoToml

OpenServerLogs

ParentModule

PositionOrRange

RebuildProcMacros

RelatedTests

ReloadWorkspace

RunFlycheck

RunTest

RunnableArgs

RunnableKind

Runnables

ServerStatusNotification

SnippetDocumentChangeOperation

Ssr

TestItemKind

TestState

ViewCrateGraph

ViewFileText

ViewHir

ViewItemTree

ViewMir

ViewRecursiveMemoryLayout

ViewSyntaxTree

WorkspaceSymbol

WorkspaceSymbolSearchKind

WorkspaceSymbolSearchScope